Execute regular queries asynchronously
We try this the easy way - send the query as it arrives then just wait
to check the result until the next query arrives. That means that the
receiving database can go process the update/insert/whatever while we go
read and parse the next item from the replication slot.

Only on commit do we actually wait for the command to fully complete
before we confirm it to the walsender as received.