Implement transaction batching
Using -b we can batch multiple transactions on the origin before we
commit them on the target server. This also includes not updating the
current position as often as before.

Note that when doing this things may be left uncommitted on the standby
if the origin suddenly stops sending data. In that case, it's
recommended to restart pg_trivialreplay with a batch size of 1 to catch
up.